+++ index.html
@@ -231,6 +231,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="js\main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 </html>
